Added test for command processor not defined. Moved test files under Command. Added DisplayWall command.
