+++ SocialNetwork/SocialNetwork.UnitTests/Command/CommandDispatcherShould.cs
@@ -1,9 +1,8 @@
-﻿using System;
-using Microsoft.VisualStudio.TestTools.UnitTesting;
+﻿using Microsoft.VisualStudio.TestTools.UnitTesting;
 using Moq;
 using SocialNetwork.Command;
 
-namespace SocialNetwork.UnitTests
+namespace SocialNetwork.UnitTests.Command
 {
     [TestClass]
     public class CommandDispatcherShould
