+++ SocialNetwork/SocialNetwork/Command/CommandTranslator.cs
@@ -15,6 +15,9 @@ namespace SocialNetwork.Command
             if (split.Length == 1)
                 return new DisplayUserTimeline(command);
 
+            if (split.Length == 2 && split[1].Equals("wall"))
+                return new DisplayWall(split[0]);
+
             if (split[1].Equals("->"))
                 return new PostMessage(split[0], command.Substring(command.IndexOf("->", StringComparison.InvariantCulture) + 3));
