+++ SocialNetwork/SocialNetwork/Command/DisplayWall.cs
@@ -1,0 +1,9 @@
+﻿namespace SocialNetwork.Command
+{
+    public class DisplayWall: UserCommand
+    {
+        public DisplayWall(string username) : base(username)
+        {
+        }
+    }
+}
