+++ SocialNetwork/SocialNetwork/Exceptions/CommandProcessorNotDefinedException.cs
@@ -1,0 +1,15 @@
+﻿using System;
+
+namespace SocialNetwork.Exceptions
+{
+    public class CommandProcessorNotDefinedException: Exception
+    {
+        public CommandProcessorNotDefinedException()
+        {
+        }
+
+        public CommandProcessorNotDefinedException(string message) : base(message)
+        {
+        }
+    }
+}
